+++ src/main/kotlin/com/blife/blife/domain/review/controller/LibBookReviewController.kt
@@ -1,10 +1,8 @@
 package com.blife.blife.domain.review.controller
 
-import com.blife.blife.domain.review.dto.BookReviewRequest
-import com.blife.blife.domain.review.dto.BookReviewResponse
 import com.blife.blife.domain.review.dto.LibBookReviewRequest
 import com.blife.blife.domain.review.dto.LibBookReviewResponse
-import com.blife.blife.domain.review.service.ReviewService
+import com.blife.blife.domain.review.service.LibBookReviewService
 import io.swagger.v3.oas.annotations.Operation
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
@@ -22,26 +20,9 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping("/review")
-class ReviewController(
-    private val reviewService: ReviewService,
+class LibBookReviewController(
+    private val libBookReviewService: LibBookReviewService,
 ) {
-    @Operation(summary = "책 리뷰 조회")
-    @GetMapping("/book/{bookId}")
-    fun getBookReviewList(
-        @PathVariable bookId: Long,
-        @PageableDefault pageable: Pageable,
-    ): ResponseEntity<Page<BookReviewResponse>> {
-        return ResponseEntity.status(HttpStatus.OK).body(reviewService.getReviewListByBookId(bookId, pageable))
-    }
-
-    @Operation(summary = "책 단일리뷰 조회")
-    @GetMapping("/book/{bookReviewId}")
-    fun getBookReview(
-        @PathVariable bookReviewId: Long,
-    ): ResponseEntity<BookReviewResponse> {
-        return ResponseEntity.status(HttpStatus.OK).body(reviewService.getReviewByBookId(bookReviewId))
-    }
-
 
     @Operation(summary = "도서관내 책 리뷰 조회")
     @GetMapping("/libBook/{libBookId}")
@@ -49,7 +30,7 @@ class ReviewController(
         @PathVariable libBookId: Long,
         @PageableDefault pageable: Pageable,
     ): ResponseEntity<Page<LibBookReviewResponse>> {
-        return ResponseEntity.status(HttpStatus.OK).body(reviewService.getReviewListByLibBookId(libBookId, pageable))
+        return ResponseEntity.status(HttpStatus.OK).body(libBookReviewService.getReviewListByLibBookId(libBookId, pageable))
     }
 
     @Operation(summary = "책 단일리뷰 조회")
@@ -57,20 +38,9 @@ class ReviewController(
     fun getBookReviewInLib(
         @PathVariable libBookReviewId: Long,
     ): ResponseEntity<LibBookReviewResponse> {
-        return ResponseEntity.status(HttpStatus.OK).body(reviewService.getBookReviewInLib(libBookReviewId))
+        return ResponseEntity.status(HttpStatus.OK).body(libBookReviewService.getBookReviewInLib(libBookReviewId))
     }
 
-    @Operation(summary = "책 리뷰 생성")
-    @PostMapping("/book/{bookId}")
-    fun createBookReview(
-        @PathVariable bookId: Long,
-        @AuthenticationPrincipal userPrincipal: UserPrincipal,
-        bookReviewRequest: BookReviewRequest,
-    ): ResponseEntity<BookReviewResponse> {
-        val userId = userPrincipal.id
-        reviewService.createBookReview(bookId, userId, bookReviewRequest)
-        return ResponseEntity.status(HttpStatus.CREATED).build()
-    }
 
     @Operation(summary = "도서관내 책 리뷰 생성")
     @PostMapping("/libBook/{libBookId}")
@@ -80,21 +50,10 @@ class ReviewController(
         libBookReviewRequest: LibBookReviewRequest,
     ): ResponseEntity<LibBookReviewResponse> {
         val userId = userPrincipal.id
-        reviewService.createBookReviewInLib(libBookId, userId, libBookReviewRequest)
+        libBookReviewService.createBookReviewInLib(libBookId, userId, libBookReviewRequest)
         return ResponseEntity.status(HttpStatus.CREATED).build()
     }
 
-    @Operation(summary = "책 리뷰 수정")
-    @PatchMapping("/book/{bookReviewId}")
-    fun updateBookReview(
-        @PathVariable bookReviewId: Long,
-        @AuthenticationPrincipal userPrincipal: UserPrincipal,
-        bookReviewRequest: BookReviewRequest,
-    ): ResponseEntity<BookReviewResponse> {
-        val userId = userPrincipal.id
-        reviewService.updateBookReview(bookReviewId, userId, bookReviewRequest)
-        return ResponseEntity.status(HttpStatus.OK).build()
-    }
 
     @Operation(summary = "도서관내 책 리뷰 수정")
     @PatchMapping("/libBook/{libBookReviewId}")
@@ -104,20 +63,10 @@ class ReviewController(
         libBookReviewRequest: LibBookReviewRequest,
     ): ResponseEntity<LibBookReviewResponse> {
         val userId = userPrincipal.id
-        reviewService.updateBookReviewInLib(libBookReviewId, userId, libBookReviewRequest)
+        libBookReviewService.updateBookReviewInLib(libBookReviewId, userId, libBookReviewRequest)
         return ResponseEntity.status(HttpStatus.OK).build()
     }
 
-    @Operation(summary = "책 리뷰 삭제")
-    @DeleteMapping("/book/{bookReviewId}")
-    fun deleteReview(
-        @PathVariable bookReviewId: Long,
-        @AuthenticationPrincipal userPrincipal: UserPrincipal,
-    ): ResponseEntity<Unit> {
-        val userId = userPrincipal.id
-        reviewService.deleteReview(bookReviewId, userId)
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build()
-    }
 
     @Operation(summary = "도서관 내 책 리뷰 삭제")
     @DeleteMapping("/libBook/{libBookReviewId}")
@@ -126,7 +75,7 @@ class ReviewController(
         @AuthenticationPrincipal userPrincipal: UserPrincipal,
     ): ResponseEntity<Unit> {
         val userId = userPrincipal.id
-        reviewService.deleteLibBookReview(libBookReviewId, userId)
+        libBookReviewService.deleteLibBookReview(libBookReviewId, userId)
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build()
     }
 }
